Make time between correct answer and next question non-blocking
Previously when the correct answer was given, the game loop paused for X seconds (blocking). This didn't allow the messages arriving after the first right answer to be parsed, which could be quite a few.
Now instead, set a global flag of "question_answered_correctly" and in the game loop, wait until a timer expires. Meanwhile while waiting, continue to process all messages as normal