Remove forced bold highlight of question. Check if a * exists in the text, first
Move the logic to the Question class. If it contains a *, don't wrap the text in stars. Will now work with upcoming MultiChoice support

Slack message API has support for a markdown disable flag in the message JSON. But the Python Slack module doesn't support setting that in the rtm_message method. It might do in the non rtm api. Boo urns